blogs tests; create/update/delete auth;

+++ part4/utils/controllers.js
@@ -23,17 +23,23 @@ const initResourceController =
       }
     });
 
-    router.delete(`/:id`, async (request, response) => {
-      await Model.findByIdAndDelete(request.params.id);
-      response.status(204).end();
-    });
+    router.delete(
+      `/:id`,
+      expressAuthMiddleware(),
+      async (request, response) => {
+        if (!request.auth.userId) return authError(response);
+        if (!(await checkUserOwn(Model, request)))
+          return forbiddenError(response);
+
+        await Model.findByIdAndDelete(request.params.id);
+        response.status(204).end();
+      },
+    );
 
     router.post(`/`, expressAuthMiddleware(), async (request, response) => {
       const body = request.body;
 
-      if (!request.auth.userId) {
-        return response.status(401).json({ error: 'token invalid' });
-      }
+      if (!request.auth.userId) return authError(response);
 
       const user = await User.findById(request.auth.userId);
 
@@ -55,8 +61,12 @@ const initResourceController =
       response.status(201).json(savedItem);
     });
 
-    router.put(`/:id`, async (request, response) => {
+    router.put(`/:id`, expressAuthMiddleware(), async (request, response) => {
       const body = request.body;
+
+      if (!request.auth.userId) return authError(response);
+      if (!(await checkUserOwn(Model, request)))
+        return forbiddenError(response);
 
       if (updateValidate) {
         const error = await updateValidate(body);
@@ -73,5 +83,17 @@ const initResourceController =
       response.json(updatedItem);
     });
   };
+
+const checkUserOwn = async (Model, request) => {
+  const item = await Model.findById(request.params.id);
+
+  return item.user && item.user.toString() === request.auth.userId;
+};
+
+const authError = (response) =>
+  response.status(401).json({ error: 'token invalid' });
+
+const forbiddenError = (response) =>
+  response.status(403).json({ error: 'forbidden' });
 
 module.exports = { initResourceController };
